Adhere to clean architecture
