+++ src/main/java/com/playground/playground/entity/NeuralNetBuilder.java
@@ -1,7 +1,8 @@
-package com.playground.playground.modelling;
+package com.playground.playground.entity;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+
 import org.deeplearning4j.nn.api.OptimizationAlgorithm;
 import org.deeplearning4j.nn.conf.Updater;
 import org.deeplearning4j.nn.weights.WeightInit;
