+++ src/main/java/com/playground/playground/interface_adapater/controller/MainController.java
@@ -1,13 +1,9 @@
-package com.playground.playground;
+package com.playground.playground.interface_adapater.controller;
 
 import java.io.IOException;
 import java.net.URL;
-import java.util.ArrayList;
 import java.util.ResourceBundle;
 
-import com.playground.playground.data.FeatureController;
-import com.playground.playground.modelling.ModelTrainingServices;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
