+++ src/main/resources/com/playground/playground/data-attributes-view.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<VBox prefHeight="426.0" prefWidth="198.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.playground.playground.DataAttributesController">
+<VBox prefHeight="426.0" prefWidth="198.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.playground.playground.interface_adapater.controller.DataAttributesController">
     <VBox spacing="10">
         <Text strokeType="OUTSIDE" strokeWidth="0.0" styleClass="data-text" text="Data" textAlignment="LEFT" translateX="75.0" wrappingWidth="45.79052734375">
          <VBox.margin>
