+++ src/test/java/com/playground/playground/data/DataTest.java
@@ -7,6 +7,12 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import com.playground.playground.entity.DatasetGenerator;
+import com.playground.playground.usecase.datasets.CircularDatasetGenerator;
+import com.playground.playground.usecase.datasets.QuadrantDatasetGenerator;
+import com.playground.playground.usecase.datasets.SpiralDatasetGenerator;
+import com.playground.playground.usecase.datasets.TransformDatasets;
 import org.junit.jupiter.api.Test;
 
 /** Test class to validate the functionality of various components in the data package. */
