+++ src/test/java/com/playground/playground/data/FeatureTest.java
@@ -7,6 +7,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+
+import com.playground.playground.entity.FeatureApplier;
+import com.playground.playground.interface_adapater.controller.FeatureController;
+import com.playground.playground.usecase.features.FeatureApplierFactory;
+import com.playground.playground.usecase.features.MultiplyFeatureApplier;
+import com.playground.playground.usecase.features.SinFeatureApplier;
+import com.playground.playground.usecase.features.SquareFeatureApplier;
 import org.junit.jupiter.api.Test;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.primitives.Pair;
