+++ src/test/java/com/playground/playground/modelling/NeuralNetTest.java
@@ -1,6 +1,8 @@
 package com.playground.playground.modelling;
 
 import java.util.ArrayList;
+
+import com.playground.playground.entity.NeuralNet;
 import org.deeplearning4j.nn.api.OptimizationAlgorithm;
 import org.deeplearning4j.nn.conf.Updater;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
